Move internal files to .internal package and exclude in dokka
