+++ viewpump/src/main/java/io/github/inflationx/viewpump/internal/-InterceptorChain.kt
@@ -1,11 +1,16 @@
 @file:JvmName("-InterceptorChain")
-package io.github.inflationx.viewpump
+package io.github.inflationx.viewpump.internal
+
+import io.github.inflationx.viewpump.InflateRequest
+import io.github.inflationx.viewpump.InflateResult
+import io.github.inflationx.viewpump.Interceptor
+import io.github.inflationx.viewpump.Interceptor.Chain
 
 /**
  * A concrete interceptor chain that carries the entire interceptor chain.
  */
 internal class InterceptorChain(private val interceptors: List<Interceptor>, private val index: Int,
-    private val request: InflateRequest) : Interceptor.Chain {
+    private val request: InflateRequest) : Chain {
 
   override fun request(): InflateRequest {
     return request
@@ -17,7 +22,8 @@ internal class InterceptorChain(private val interceptors: List<Interceptor>, pri
     }
 
     // Call the next interceptor in the chain.
-    val next = InterceptorChain(interceptors, index + 1, request)
+    val next = InterceptorChain(interceptors, index + 1,
+        request)
     val interceptor = interceptors[index]
 
     return interceptor.intercept(next)
